trust that field specs come with registers
The input files are not entirely consistent, but it seems that instead of
guessing field parents after parsing everything, the parent registers for
fields can be guessed based on the most recently read register. Change
the field map construction to work like that.

The fields for a register usually come right after in the headers. There
are some exceptions which appear at least as missing registers so print
a warning for each. Not a problem yet.